+++ src/autocomplete/inline_test.js
@@ -37,6 +37,11 @@ var completions = [
     {
         snippet: "foobar2",
         score: 0
+    },
+    {
+        value: "f should not show inline",
+        score: 0,
+        hideInlinePreview: true
     }
 ];
 
@@ -220,6 +225,40 @@ module.exports = {
         editor.renderer.$loop._flush();
         assert.strictEqual(inline.isOpen(), false);
         assert.strictEqual(getAllLines(), textBase + "f");
+        done();
+    },
+    "test: should respect hideInlinePreview": function(done) {
+        // By default, this option is set to hide.
+        inline.show(editor, completions[5], "f");
+        editor.renderer.$loop._flush();
+        assert.equal(getAllLines(), textBase + "f");
+        assert.strictEqual(inline.isOpen(), false);
+        inline.hide();
+        editor.renderer.$loop._flush();
+
+        // Now it should be shown.
+        completions[5].hideInlinePreview = false;
+
+        inline.show(editor, completions[5], "f");
+        editor.renderer.$loop._flush();
+        assert.equal(getAllLines(), textBase + "f should not show inline");
+        assert.strictEqual(inline.isOpen(), true);
+        inline.hide();
+        editor.renderer.$loop._flush();
+
+        // Now it should be shown.
+        completions[5].hideInlinePreview = undefined;
+
+        inline.show(editor, completions[5], "f");
+        editor.renderer.$loop._flush();
+        assert.equal(getAllLines(), textBase + "f should not show inline");
+        assert.strictEqual(inline.isOpen(), true);
+        inline.hide();
+        editor.renderer.$loop._flush();
+
+        // Reset to state before test.
+        completions[5].hideInlinePreview = true;
+
         done();
     },
     tearDown: function() {