+++ src/config.js
@@ -223,6 +223,6 @@ var reportErrorIfPathIsNotConfigured = function() {
     }
 };
 
-exports.version = "1.26.0";
+exports.version = "1.28.0";